// workspaces/P20260914-1728-网页版票夹管理发票/03_engineering/scaffold/web/src/api/generated/model/invoiceView.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface InvoiceView {
  amount?: string;
  attachmentCount?: number;
  category?: string;
  createdAt?: string;
  id?: string;
  invoiceCode?: string;
  invoiceNumber?: string;
  issuedDate?: string;
  medium?: string;
  remark?: string;
  status?: string;
  taxAmount?: string;
  title?: string;
  totalAmount?: string;
  updatedAt?: string;
}